requirements.txt
aspose-slides==25.8.1